requirements.txt
dash
ipykernel
matplotlib 
numpy
pandas
psycopg2
python-dotenv
requests
scikit-learn
seaborn
snowflake-connector-python
sqlalchemy